+++ jwst/extract_2d/grisms.py
@@ -476,10 +476,7 @@ def extract_grism_objects(input_model,
                 # The overall subarray offset is recorded in model.meta.subarray.
                 # nslit = obj.sid - 1  # catalog id starts at zero
                 new_slit.name = "{0}".format(obj.sid)
-                if obj.is_star:
-                    new_slit.is_star = True
-                else:
-                    new_slit.source_type = 'EXTENDED'
+                new_slit.is_star = obj.is_star
                 new_slit.xstart = xmin + 1  # fits pixels
                 new_slit.xsize = ext_data.shape[1]
                 new_slit.ystart = ymin + 1  # fits pixels